requirements.txt
Flask
pyyaml
stable-baselines3==2.0.0
pygame 
matplotlib 
numpy 
gym
shimmy>=0.2.1